+++ app/src/main/java/code/yakovenko/tariffka/domain/usecase/delete/DeleteSupportTicketUseCase.kt
@@ -1,11 +1,21 @@
 package code.yakovenko.tariffka.domain.usecase.delete
 
+import code.yakovenko.tariffka.domain.exception.SupportTicketNotFoundException
 import code.yakovenko.tariffka.domain.repository.SupportTicketRepository
+import kotlinx.coroutines.flow.firstOrNull
+import javax.inject.Inject
+import kotlin.uuid.ExperimentalUuidApi
+import kotlin.uuid.Uuid
 
-class DeleteSupportTicketUseCase(
+@OptIn(ExperimentalUuidApi::class)
+class DeleteSupportTicketUseCase @Inject constructor(
     private val supportTicketRepository: SupportTicketRepository
 ) {
-    suspend operator fun invoke(supportTicketId: Int) {
+    suspend operator fun invoke(supportTicketId: Uuid) {
+        if (supportTicketRepository.readById(supportTicketId).firstOrNull() == null) {
+            throw SupportTicketNotFoundException(supportTicketId)
+        }
+
         supportTicketRepository.deleteById(supportTicketId)
     }
 }
